Move useDocumentCaption into hooks folder.
